Optimisations and code cleanup - Added small date time and battery poll to update during active screensaver - Cleaned up module functions to use void return where no value was needed - Fixed allocated memory not being freed when content fails to load - Fixed crash dialogue box not finish animating and staying off screen - Fixed several small memory leaks with function calls - Improved collection, history, and tag detection speed for content - Reduced memory usage for JSON reading during directory scanning - Renamed unused nothing parameter to auto_assign for consistency

+++ module/muxthemeopt.c
@@ -91,7 +91,7 @@ static void restore_width_dropdown(lv_obj_t *dropdown, int16_t stored) {
 static void save_width_dropdown(lv_obj_t *dropdown, int original, const char *key, int16_t *cfg, int *is_modified) {
     uint32_t sel = lv_dropdown_get_selected(dropdown);
     if ((int) sel == original) return;
-    int16_t val = (sel >= 1) ? (int16_t) (sel + 9) : 0;
+    int16_t val = (int16_t) ((sel >= 1) ? (int) sel + 9 : 0);
     write_text_to_file(key, "w", INT, (int) val);
     *cfg = val;
     (*is_modified)++;
